+++ libV2ray/libV2ray/libV2ray_wrapper.go
@@ -29,7 +29,6 @@ const (
 	coreAsset = "v2ray.location.asset"
 )
 
-// CoreController manages v2fly core instance lifecycle
 type CoreController struct {
 	CallbackHandler CoreCallbackHandler
 	statsManager    corestats.Manager
@@ -38,36 +37,27 @@ type CoreController struct {
 	IsRunning       bool
 }
 
-// CoreCallbackHandler defines interface for receiving callbacks from the core service
 type CoreCallbackHandler interface {
 	Startup() int
 	Shutdown() int
 	OnEmitStatus(int, string) int
 }
 
-// consoleLogWriter implements a log writer without datetime stamps
-// (Android system already adds timestamps to each log line)
 type consoleLogWriter struct {
-	logger *log.Logger // Standard logger
+	logger *log.Logger
 }
 
-// setEnvVariable safely sets an environment variable and logs any errors encountered.
 func setEnvVariable(key, value string) {
 	if err := os.Setenv(key, value); err != nil {
 		log.Printf("Failed to set environment variable %s: %v. Please check your configuration.", key, err)
 	}
 }
 
-// InitCoreEnv initializes environment variables and file system handlers for the core
-// It sets up asset path, certificate path, XUDP base key and customizes the file reader
-// to support Android asset system
 func InitCoreEnv(envPath string, key string) {
-	// Set asset/cert paths
 	if len(envPath) > 0 {
 		setEnvVariable(coreAsset, envPath)
 	}
 
-	// Custom file reader with path validation
 	corefilesystem.NewFileReader = func(path string) (io.ReadCloser, error) {
 		if _, err := os.Stat(path); os.IsNotExist(err) {
 			_, file := filepath.Split(path)
@@ -77,10 +67,7 @@ func InitCoreEnv(envPath string, key string) {
 	}
 }
 
-// NewCoreController initializes and returns a new CoreController instance
-// Sets up the console log handler and associates it with the provided callback handler
 func NewCoreController(s CoreCallbackHandler) *CoreController {
-	// Register custom logger
 	if err := coreapplog.RegisterHandlerCreator(
 		coreapplog.LogType_Console,
 		func(lt coreapplog.LogType, options coreapplog.HandlerCreatorOptions) (corecommlog.Handler, error) {
@@ -95,11 +82,6 @@ func NewCoreController(s CoreCallbackHandler) *CoreController {
 	}
 }
 
-
-
-// StartLoop initializes and starts the core processing loop
-// Thread-safe method that configures and runs the v2fly core with the provided configuration
-// Returns immediately if the core is already running
 func (x *CoreController) StartLoop(configContent string) (err error) {
 	x.coreMutex.Lock()
 	defer x.coreMutex.Unlock()
@@ -112,8 +94,6 @@ func (x *CoreController) StartLoop(configContent string) (err error) {
 	return x.doStartLoop(configContent)
 }
 
-// StopLoop safely stops the core processing loop and releases resources
-// Thread-safe method that shuts down the core instance and triggers necessary callbacks
 func (x *CoreController) StopLoop() error {
 	x.coreMutex.Lock()
 	defer x.coreMutex.Unlock()
@@ -125,7 +105,6 @@ func (x *CoreController) StopLoop() error {
 	return nil
 }
 
-// doShutdown shuts down the v2fly instance and cleans up resources
 func (x *CoreController) doShutdown() {
 	if x.coreInstance != nil {
 		if err := x.coreInstance.Close(); err != nil {
@@ -137,7 +116,6 @@ func (x *CoreController) doShutdown() {
 	x.statsManager = nil
 }
 
-// doStartLoop sets up and starts the v2fly core
 func (x *CoreController) doStartLoop(configContent string) error {
 	log.Println("Initializing core...")
 	config, err := coreserial.LoadJSONConfig(strings.NewReader(configContent))
@@ -165,7 +143,6 @@ func (x *CoreController) doStartLoop(configContent string) error {
 	return nil
 }
 
-// measureInstDelay measures the delay for an instance to a given URL
 func measureInstDelay(ctx context.Context, inst *core.Instance, url string) (int64, error) {
 	if inst == nil {
 		return -1, errors.New("core instance is nil")
@@ -206,7 +183,6 @@ func measureInstDelay(ctx context.Context, inst *core.Instance, url string) (int
 	return time.Since(start).Milliseconds(), nil
 }
 
-// Log writer implementation
 func (w *consoleLogWriter) Write(s string) error {
 	w.logger.Print(s)
 	return nil
@@ -216,11 +192,43 @@ func (w *consoleLogWriter) Close() error {
 	return nil
 }
 
-// createStdoutLogWriter creates a logger that won't print date/time stamps
 func createStdoutLogWriter() corecommlog.WriterCreator {
 	return func() corecommlog.Writer {
 		return &consoleLogWriter{
 			logger: log.New(os.Stdout, "", 0),
 		}
 	}
+}
+
+func GetFreePorts(count int) ([]int, error) {
+	var ports []int
+	for range count {
+		addr, err := net.ResolveTCPAddr("tcp", "localhost:0")
+		if err != nil {
+			return ports, err
+		}
+
+		l, err := net.ListenTCP("tcp", addr)
+		if err != nil {
+			return ports, err
+		}
+		ports = append(ports, l.Addr().(*net.TCPAddr).Port)
+		l.Close()
+	}
+	return ports, nil
+}
+
+func GetFreePort() (int, error) {
+	addr, err := net.ResolveTCPAddr("tcp", "localhost:0")
+	if err != nil {
+		return 0, err
+	}
+
+	l, err := net.ListenTCP("tcp", addr)
+	if err != nil {
+		return 0, err
+	}
+	port := l.Addr().(*net.TCPAddr).Port
+	l.Close()
+	return port, nil
 }
