--- libV2ray/libV2ray/ping.go
@@ -1,43 +1,0 @@
-package libV2ray
-
-import (
-	"context"
-	"fmt"
-	"strings"
-	"time"
-
-	core "github.com/v2fly/v2ray-core/v5"
-	coreserial "github.com/v2fly/v2ray-core/v5/infra/conf/serial"
-)
-
-// MeasureDelay measures network latency to a specified URL through the current core instance
-// Uses a 12-second timeout context and returns the round-trip time in milliseconds
-// An error is returned if the connection fails or returns an unexpected status
-func (x *CoreController) MeasureDelay(url string) (int64, error) {
-	ctx, cancel := context.WithTimeout(context.Background(), 12*time.Second)
-	defer cancel()
-
-	return measureInstDelay(ctx, x.coreInstance, url)
-}
-
-// MeasureOutboundDelay measures the outbound delay for a given configuration and URL
-func MeasureOutboundDelay(ConfigureFileContent string, url string) (int64, error) {
-	config, err := coreserial.LoadJSONConfig(strings.NewReader(ConfigureFileContent))
-	if err != nil {
-		return -1, fmt.Errorf("Configuration load error: %w", err)
-	}
-
-	// Simplify config for testing
-	config.Inbound = nil
-
-	inst, err := core.New(config)
-	if err != nil {
-		return -1, fmt.Errorf("Instance creation failed: %w", err)
-	}
-
-	if err := inst.Start(); err != nil {
-		return -1, fmt.Errorf("startup failed: %w", err)
-	}
-	defer inst.Close()
-	return measureInstDelay(context.Background(), inst, url)
-}
