--- libV2ray/libV2ray/stats.go
@@ -1,17 +1,0 @@
-package libV2ray
-
-import "fmt"
-
-// QueryStats retrieves and resets traffic statistics for a specific outbound tag and direction
-// Returns the accumulated traffic value and resets the counter to zero
-// Returns 0 if the stats manager is not initialized or the counter doesn't exist
-func (x *CoreController) QueryStats(tag string, direct string) int64 {
-	if x.statsManager == nil {
-		return 0
-	}
-	counter := x.statsManager.GetCounter(fmt.Sprintf("outbound>>>%s>>>traffic>>>%s", tag, direct))
-	if counter == nil {
-		return 0
-	}
-	return counter.Set(0)
-}
